# Objective
`prepare_uniform_components`'s commands must be run with exclusive access to the render world and can take quite a bit of time for components on lots of entities, particularly with archetypes with many big components. This is doing redundant work that is already being done in `Extract`.

## Solution
 - Implement `Default` on `DynamicUniformIndex`.
 - Move adding `DynamicUniformIndex` into Extract instead of `Prepare`.
 - Change `prepare_uniform_components` to query for `&mut DynamicUniformIndex` instead of using commands.

## Performance
This was tested against the `many_cubes` stress test. Here are the respective timing changes:

|stage/system|main|this PR|
|:--|:--|:--|
|Full Frame|20.52ms|18.79ms|
|Extract (Stage)|3.7ms|3.83ms|
|Prepare (Stage)|2.64ms|1.13ms|
|extract_meshes (system)|1.45ms|1.69us|
|extract_meshes (commands)|881.75us|1.48ms|
|prepare_uniform_components<MeshUniform> (system)|1.12ms|929.56us|
|prepare_uniform_components<MeshUniform> (commands)|1.26ms|253ns|

---

## Changelog
TODO

## Migration Guide
TODO