@james7132 2022-06-19
Wanted to just note that if we merge #4902, we can avoid the secondary copy inside the extraction commands by adding a `ExtractedComponentUniforms<C>` resource we write to instead of inserting via `Commands`. This does make it harder to query for the the uniforms after the fact, but this minimizes the number of heavy copies being done. This effectively forces the extract stage systems to do what `UniformComponentPlugin` already did before, but it definitely avoids heavier copies, both into the render world and within it.

I also tested to see if we could defer the `inverse().transpose()` computation in `extract_meshes`, and it does seem like it shuffles about 100us of time in `many_cubes` to the prepare stage. Unlike before, this can be done in-place, without any copying, only a temporary on the stack is used.