@superdump 2022-06-19
> Wanted to just note that if we merge #4902, we can avoid the secondary copy inside the extraction commands by adding a `ExtractedComponentUniforms<C>` resource we write to instead of inserting via `Commands`. This does make it harder to query for the the uniforms after the fact, but this minimizes the number of heavy copies being done. This effectively forces the extract stage systems to do what `UniformComponentPlugin` already did before, but it definitely avoids heavier copies, both into the render world and within it.
> 
> EDIT: Tried this, the `queue_material_meshes` system is dependent on `MeshUniform` as a component.
> 
> I also tested to see if we could defer the `inverse().transpose()` computation in `extract_meshes`, and it does seem like it shuffles about 100us of time in `many_cubes` to the prepare stage. Unlike before, this can be done in-place, without any copying, only a temporary on the stack is used.

This again becomes a question of whether the model matrix is used/desired to be used elsewhere in the render schedule such that not calculating it upfront incurs calculation of it multiple times. If it just moves time from one place to another with no other overall performance benefits then the only pro is that it makes the extract stage shorter. That would be enough but only if we think no one ever needs the model matrix. I wonder if TAA would need it for motion vectors or how that works…