@james7132 2022-06-19
> Could you test and profile it to see if it does make a practical performance difference? It would be a win for correctness in case something doesn’t actually ever get set.

Tried this with a `if let Some(index) = dynamic.index()`. The Render phase perf doesn't tangibly change (7.49ms -> 7.50ms), though I think this may cause incorrect bindings if we don't panic on not finding an index. Changing it to panic via `unwrap` negatively affects render perf on my machine (7.49ms ->7.63ms).