@superdump 2022-06-19
> > Could you test and profile it to see if it does make a practical performance difference? It would be a win for correctness in case something doesn’t actually ever get set.
> 
> Tried this with a `if let Some(index) = dynamic.index()`. The Render phase perf doesn't tangibly change (7.49ms -> 7.50ms), though I think this may cause incorrect bindings if we don't panic on not finding an index. Changing it to panic via `unwrap` negatively affects render perf on my machine (7.49ms ->7.63ms).

Good to know that panic incurs that performance hit. I was thinking that a missing index would somehow cause that entity not to be drawn by propagating up an error or something. But if we don’t already have error returns from draw functions then maybe it’s not worth it. I’m just kind of expecting it to be easy enough to make code where some entities never have their dynamic index updated and then they will be drawn using whatever the transform is for index 0. I suppose another way to handle it would be to make that model matrix produce vertices containing nans in the clip position and then it will be dropped, but that feels like a hack where it would be better to just not draw the thing.