@superdump 2022-06-21
> > Good to know that panic incurs that performance hit. I was thinking that a missing index would somehow cause that entity not to be drawn by propagating up an error or something. But if we don’t already have error returns from draw functions then maybe it’s not worth it. I’m just kind of expecting it to be easy enough to make code where some entities never have their dynamic index updated and then they will be drawn using whatever the transform is for index 0. I suppose another way to handle it would be to make that model matrix produce vertices containing nans in the clip position and then it will be dropped, but that feels like a hack where it would be better to just not draw the thing.
> 
> Having seen the perf hit, I tried the opposite and changed `RenderCommand` to return a `Result<(), RenderCommandError>` instead and changed the existing `unwraps` into `ok_or(...)?`s, and saw a slight improvement in perf (7.49ms -> 7.37ms in the same stress test). This is likely because we're removing all machinery for panicking from the functions, allowing for fewer instructions/smaller jumps. I can either add it to this PR, though I'd rather not due to how large/controversial it would be relative to this change.

Sounds reasonable to do it in a separate PR. If you don't intend to do that straight away, could you add a TODO comment?